* Disables MMXEXT pixelformat conversions, which are not good * MMXEXT optimized YUV-> RGB24 * Fixed routine lookup for better benchmarks
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@1461 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gavl/include/mix.h
@@ -42,7 +42,7 @@ typedef struct gavl_mix_input_channel_s
   int index; /* Which input channel */
   union      /* Weighing factor     */
     {
-    float    f_float;
+    double    f_float;
     int8_t   f_8;
     int16_t  f_16;
     int32_t  f_32;
